+++ changelog.md
@@ -10,10 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ ] refuse from stateful `Imohash` class in favor of two functions with arguments instead
 - [ ] add support for `typing.IO` objects instead of too narrow file path
 - [ ] move multithreading logic from Python to Rust
+
+## [0.2.0] — 2025-05-05
+
 - [x] cover with unit-tests
+- [x] feature: add special `-` argument for reading `file_path` list from STDIN for CLI application
 
 ## [0.1.0] — 2025-04-16
 
 Initial commit
 
+[0.2.0]: https://github.com/unsektor/py-imohash/releases/tag/0.2.0
 [0.1.0]: https://github.com/unsektor/py-imohash/releases/tag/0.1.0